Extract code generation logic into its own module
